app/services/embeddings.py
import os
import google.generativeai as genai
from app.core.config import settings

genai.configure(api_key=settings.GEMINI_API_KEY)

def get_embedding(text: str) -> list[float]:
    """將文字轉為 768 維向量 (Gemini 預設維度)"""
    result = genai.embed_content(
        model="models/text-embedding-004",
        content=text,
        task_type="retrieval_document",
        title="JackMa Memory"
    )
    return result['embedding']
