requirements.txt
requests>=2.31.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
openai>=1.12.0
httpx>=0.26.0
aiofiles>=23.2.1
python-dotenv>=1.0.0
elevenlabs>=1.0.0
pydantic-settings>=2.2.1
google-generativeai>=0.7.2
google-genai>=1.0.0
google-cloud-storage>=2.14.0
sqlalchemy>=2.0.25
pgvector>=0.3.6
psycopg2-binary>=2.9.9

# Auth
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0

# Image processing
Pillow>=10.0.0

# Rate Limiting
slowapi>=0.1.9

# 繁簡中文轉換（TTS 用簡體發音，顯示用繁體）
zhconv>=1.4.3

# LiveKit（全部鎖 1.5.1，移除 minimax plugin 用自訂 wrapper）
livekit>=1.0.0,<2.0.0
livekit-agents==1.5.1
livekit-plugins-google==1.5.1
livekit-plugins-elevenlabs==1.5.1
livekit-plugins-silero==1.5.1
livekit-plugins-anthropic==1.5.1
livekit-plugins-deepgram==1.5.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
